performance and logging additions
created the .update() ThermostatApp function that pulls down the json description of the entire project and the state of all widgets within it, and now we use that at the beginning of every loop to update the entire application state within a single call.

I'm also sending the collected cache_hits and cache_miss stats to splunk, in order to monitor performance. As of now, almost all application state comes from the single .update() call, and the only cache misses I'm seeing I believe can be attributed to the dht22_read() method lagging as the sensor flakes out a bit (as they are prone to do).